+++ src/corpus/sources/SourcesPanel.vue
@@ -10,12 +10,15 @@ import ActionButton from "@/components/ActionButton.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 import useLocale from "@/i18n/locale.composable";
 import MaxHeight from "@/components/MaxHeight.vue";
-import type { ProgressHandler } from "@/api/api.types";
+import type { FileMeta, ProgressHandler } from "@/api/api.types";
 import useMessenger from "@/message/messenger.composable";
 import { getFilenameExtension } from "@/util";
 import { FORMATS_EXT, type FileFormat } from "@/api/corpusConfig";
 import FileUpload from "@/components/FileUpload.vue";
 import { useAuth } from "@/auth/auth.composable";
+import SortableTable, {
+  type SortableTableColumn,
+} from "@/components/SortableTable.vue";
 
 const props = defineProps<{
   corpusId: string;
@@ -32,7 +35,7 @@ const {
 } = useCorpus(props.corpusId);
 const { filesize } = useLocale();
 const { alert, alertError } = useMessenger();
-const { t } = useI18n();
+const { t, locale } = useI18n();
 const { canWrite } = useAuth();
 
 const info = computedAsync(getInfo);
@@ -40,6 +43,26 @@ const totalSize = computed(() =>
   sources.value.reduce((sum, source) => sum + Number(source.size), 0),
 );
 const accept = computed(() => extensions.value.map((ext) => `.${ext}`).join());
+
+const columns = computed<SortableTableColumn<FileMeta>[]>(() => {
+  const cols: SortableTableColumn<FileMeta>[] = [
+    {
+      title: t("filename"),
+      thClass: "w-full",
+      comparator: (a, b) => a.name.localeCompare(b.name, locale.value),
+    },
+    {
+      title: t("fileSize"),
+      thClass: "text-end",
+      comparator: (a, b) => a.size - b.size,
+    },
+  ];
+
+  if (canWrite("corpora", props.corpusId))
+    cols.push({ title: t("file.operations"), thClass: "sr-only" });
+
+  return cols;
+});
 
 async function fileHandler(files: File[], onProgress: ProgressHandler) {
   if (!files[0]) throw new RangeError("No files");
@@ -77,40 +100,35 @@ async function fileHandler(files: File[], onProgress: ProgressHandler) {
   </div>
   <MaxHeight :max-height="400">
     <PendingContent :on="`corpus/${corpusId}/sources/list`">
-      <table v-if="sources.length" class="w-full mt-4 striped">
-        <thead>
-          <tr>
-            <th class="w-full">{{ $t("filename") }}</th>
-            <th class="text-right">{{ $t("fileSize") }}</th>
-            <th v-if="canWrite('corpora', corpusId)" class="sr-only">
-              {{ $t("file.operations") }}
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr v-for="source in sources" :key="source.path">
-            <td>
-              <router-link
-                :to="`/library/corpus/${corpusId}/sources/${source.name}`"
-              >
-                {{ source.name }}
-              </router-link>
-            </td>
-            <td class="text-right whitespace-nowrap">
-              {{ filesize(source.size) }}
-            </td>
-            <td v-if="canWrite('corpora', corpusId)" class="text-right">
-              <ActionButton
-                class="hover:button-danger button-slim text-sm"
-                @click="deleteSource(source).catch(alertError)"
-              >
-                <PhTrash class="inline mb-0.5" />
-                <span class="sr-only">{{ $t("delete") }}</span>
-              </ActionButton>
-            </td>
-          </tr>
-        </tbody>
-      </table>
+      <SortableTable
+        v-if="sources.length"
+        :columns
+        :rows="sources"
+        :get-row-key="(source) => source.path"
+        class="w-full mt-4 striped"
+      >
+        <template v-slot="{ row: source }">
+          <td>
+            <router-link
+              :to="`/library/corpus/${corpusId}/sources/${source.name}`"
+            >
+              {{ source.name }}
+            </router-link>
+          </td>
+          <td class="text-end whitespace-nowrap">
+            {{ filesize(source.size) }}
+          </td>
+          <td v-if="canWrite('corpora', corpusId)">
+            <ActionButton
+              class="hover:button-danger button-slim text-sm"
+              @click="deleteSource(source).catch(alertError)"
+            >
+              <PhTrash class="inline mb-0.5" />
+              <span class="sr-only">{{ $t("delete") }}</span>
+            </ActionButton>
+          </td>
+        </template>
+      </SortableTable>
     </PendingContent>
   </MaxHeight>