+++ CHANGELOG.md
@@ -15,6 +15,7 @@ As this project is a user-facing application, the places in the semantic version
 
 ### Added
 
+- Sort texts by filename/size [#46](https://github.com/spraakbanken/mink-frontend/issues/46)
 - View result file [#170](https://github.com/spraakbanken/mink-frontend/issues/170)
 - Explain export file types [#192](https://github.com/spraakbanken/mink-frontend/issues/192)
 - List allowed source file extensions on upload area [#234](https://github.com/spraakbanken/mink-frontend/issues/234)